+++ frontend/src/pages/order/Basket.js
@@ -1,6 +1,10 @@
 function Basket(){
     return(
-        <h2>장바구니 입니다.</h2>
+        <h2>SHOPPING CART</h2>
+
+        
+
+
     )
 }
 export default Basket;
