+++ src/models/Product.js
@@ -1,5 +1,7 @@
 import mongoose from "mongoose";
 
+//아이템 선언
+
 const productSchema = new mongoose.Schema({
     proName: { type: String, required: true, unique: true },
     proKindName: { type: String, required: true },
